requirements.txt
glymur
flask
google-cloud-storage
gunicorn
